Refactor entity: remove LoveType FK, add report entities

+++ src/main/java/mandarin/com/mandarin_backend/controller/ReportController.java
@@ -41,7 +41,7 @@ public class ReportController {
      */
     @GetMapping("/chat/{characterId}")
     public ResponseEntity<ApiResponse<ChatReportResponseDto>> getChatReport(
-            @PathVariable Long characterId) {
+            @PathVariable Integer characterId) {
 
         ApiResponse<ChatReportResponseDto> response = reportService.getChatReport(characterId);
